pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "procawake"
version = "0.1.0"
description = "Process-aware Windows sleep prevention — keep your screen alive when specific apps are running"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Uwais Iqbal" }]
keywords = ["windows", "sleep", "caffeinate", "awake", "power-management", "system-tray", "keep-awake"]
classifiers = [
    "Operating System :: Microsoft :: Windows",
    "Environment :: Win32 (MS Windows)",
    "Topic :: System :: Power (UPS)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "psutil>=5.9",
    "pystray>=0.19",
    "Pillow>=10.0",
    "tomli_w>=1.0",
]

[project.scripts]
procawake = "procawake.cli:main"

[project.gui-scripts]
procawake-tray = "procawake.app:main"

[project.urls]
Homepage = "https://github.com/uwais/procawake"
Issues = "https://github.com/uwais/procawake/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/procawake"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
